+++ misc/update_DESCRIPTION_NEWS.R
@@ -373,6 +373,13 @@ my_news$add_subtitle("Bugfixes")
 my_news$add_bullet(c("changing error messages"))
 
 
+# fixing plugin -----------------------------------------------------------
+
+my_desc$bump_version("patch")
+my_news$add_version(my_desc$get_version())
+my_news$add_subtitle("Bugfixes")
+my_news$add_bullet(c("fixing plugin for coment line setting"))
+
 # save everything ---------------------------------------------------------
 
 my_desc$set("Date", Sys.Date())